Split RequestLoggingFilter into an additional RequestLoggingFilterBase
* The RequestLoggingFilterBase has no dependencies to
  com.sap.hcp.cf.logging.servlet.dynlog
* For scenarios without the requirement to dynamically set log levels,
  the following dependencies can then be omitted:
  - jaxb-api
  - jaxb-runtime
  - java-jwt
* Java doc needs review
